TrainBot-Gradio/requirements.txt
google-generativeai==0.8.5
gradio==5.49.1
gradio_client==1.13.3
openai==2.8.1
python-dotenv==1.2.1